refactor: hooks/useScriptExecution
